usecases/autoclaim-insurance/wxo-fileupload/requirements.txt
fastapi
python-multipart
ibm-cos-sdk
python-dotenv
uvicorn
python-magic
standard-imghdr